requirements.txt
python-docx
sklearn
nltk
pandas
numpy
